moved toIsoDateString helper function to common area and imported into relevant files
tidied imports
